refactor: lazy load outbound transcript mirroring
