+++ src/config/sessions/transcript.runtime.ts
@@ -1,0 +1,1 @@
+export { appendAssistantMessageToSessionTranscript } from "./transcript.js";
